arquitecture refactoring

+++ src/main/java/com/ochre/music/product/write/ProductWriteRepository.java
@@ -4,6 +4,8 @@ import com.ochre.music.product.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ProductWriteRepository extends JpaRepository<ProductEntity, Long> {
+public interface ProductWriteRepository extends JpaRepository<ProductEntity, BigInteger> {
 }
